+++ CHANGELOG.md
@@ -9,7 +9,10 @@ and this project adheres to
 ### Added
 
 - @cosmjs/cosmwasm-stargate: Add `SigningCosmWasmClient.instantiate2` ([#1407]).
+- @cosmjs/stargate: `IndexedTx` and `DeliverTxResponse` now have a
+  `msgResponses` field ([#1305]).
 
+[#1305]: https://github.com/cosmos/cosmjs/issues/1305
 [#1407]: https://github.com/cosmos/cosmjs/pull/1407
 
 ### Changed